+++ lib/savon/header.rb
@@ -10,6 +10,7 @@ module Savon
 
       @wsse_auth      = locals[:wsse_auth].nil? ? globals[:wsse_auth] : locals[:wsse_auth]
       @wsse_timestamp = locals[:wsse_timestamp].nil? ? globals[:wsse_timestamp] : locals[:wsse_timestamp]
+      @wsse_signature = locals[:wsse_signature].nil? ? globals[:wsse_signature] : locals[:wsse_signature]
 
       @global_header  = globals[:soap_header]
       @local_header   = locals[:soap_header]
@@ -21,7 +22,7 @@ module Savon
     end
 
     attr_reader :local_header, :global_header, :gyoku_options,
-                :wsse_auth, :wsse_timestamp
+                :wsse_auth, :wsse_timestamp, :wsse_signature
 
     def empty?
       @header.empty?
@@ -76,6 +77,10 @@ module Savon
       wsse = Akami.wsse
       wsse.credentials(*wsse_auth) if wsse_auth
       wsse.timestamp = wsse_timestamp if wsse_timestamp
+      if wsse_signature && wsse_signature.have_document?
+        wsse.signature = wsse_signature
+      end
+
       wsse
     end